Refactoring UserSolutionRepositoryTest in only 1 class BUT detached HEAD
