+++ itachallenge-user/src/test/java/com/itachallenge/user/repository/UserSolutionRepositoryTest.java
@@ -21,8 +21,9 @@ import reactor.test.StepVerifier;
 
 import java.io.IOException;
 import java.io.InputStream;
-import java.time.Duration;
 import java.util.*;
+
+import static org.junit.Assert.assertNotNull;
 import static org.junit.jupiter.api.Assertions.*;
 
 import static com.jayway.jsonpath.internal.path.PathCompiler.fail;
@@ -35,7 +36,7 @@ public class UserSolutionRepositoryTest {
 
     @Container
     static MongoDBContainer container = new MongoDBContainer("mongo")
-            .withStartupTimeout(Duration.ofSeconds(60));
+            .withExposedPorts(27017);
 
     @DynamicPropertySource
     static void initMongoProperties(DynamicPropertyRegistry registry) {
@@ -49,7 +50,6 @@ public class UserSolutionRepositoryTest {
     private IUserSolutionRepository userSolutionRepository;
 
     private static List<UserSolutionDocument> userSolutions;
-
     private static UUID testUuid;
     private static UUID testUserUuid;
     private static UUID testChallengeUuid;
@@ -59,6 +59,16 @@ public class UserSolutionRepositoryTest {
     private static int score;
     private static int all;
     private static UUID testSolutionDocumentUuid;
+
+    UUID uuid_1 = UUID.fromString("8ecbfe54-fec8-11ed-be56-0242ac120001");
+    UUID uuid_2 = UUID.fromString("26977eee-89f8-11ec-a8a3-0242ac120002");
+    UUID uuid_3 = UUID.fromString("26977eee-89f8-11ec-a8a3-0242ac120003");
+    UUID userId1 = UUID.fromString("442b8e6e-5d57-4d12-9be2-3ff4f26e7d79");
+    UUID userId2 = UUID.fromString("26977eee-89f8-11ec-a8a3-0242ac120003");
+
+    private int scoreZmq;
+    private String errorsZmq;
+    private UserSolutionDocument userSolutionZmq;
 
     @BeforeAll
     public static void deserializeJson() throws IOException {
@@ -112,14 +122,53 @@ public class UserSolutionRepositoryTest {
 
             userSolutions.add(userSolution);
         }
+
+        all += 3;
+        allBookmarked += 3;
+        /* 16/10/2024 Michel: Para que cuadren los tests al refactorizar todos en 1 sola clase
+        ** para que no colisione el desarrollo de Aebal con el de Dario */
     }
 
     @BeforeEach
     void setup(){
 
-        userSolutionRepository.deleteAll();
+        userSolutionRepository.deleteAll().block();
 
         userSolutionRepository.saveAll(Flux.fromIterable(userSolutions)).blockLast();
+
+        UUID challengeId1 = UUID.fromString("7fc6a737-dc36-4e1b-87f3-120d81c548aa");
+        UUID challengeId2 = UUID.fromString("5c1a97e5-1cca-4144-9981-2de1fb73b178");
+
+        UUID languageId1 = UUID.fromString("1e047ea2-b787-49e7-acea-d79e92be3909");
+        UUID languageId2 = UUID.fromString("09fabe32-7362-4bfb-ac05-b7bf854c6e0f");
+
+        UUID solutionId1 = UUID.fromString("1e047ea2-b787-49e7-acea-d79e92be3909");
+        UUID solutionId2 = UUID.fromString("09fabe32-7362-4bfb-ac05-b7bf854c6e0f");
+
+        String solutionText1 = "Ipsum.. 1";
+        String solutionText2 = "Ipsum.. 2";
+
+        SolutionDocument solutionDocument1 = new SolutionDocument(solutionId1, solutionText1);
+        SolutionDocument solutionDocument2 = new SolutionDocument(solutionId2, solutionText2);
+
+        List<SolutionDocument> solutionDocumentList = List.of(solutionDocument1, solutionDocument2);
+
+        UserSolutionDocument userScoreDocument1 = new UserSolutionDocument(uuid_1, userId1, challengeId1, languageId1, true, ChallengeStatus.STARTED, 90, "xxx", solutionDocumentList);
+        UserSolutionDocument userScoreDocument2 = new UserSolutionDocument(uuid_2, userId2, challengeId2, languageId2, true, ChallengeStatus.STARTED, 90, "xxx", solutionDocumentList);
+        UserSolutionDocument userScoreDocument3 = new UserSolutionDocument(uuid_3, userId1, challengeId1, languageId1, true, ChallengeStatus.STARTED, 90, "xxx", solutionDocumentList);
+
+        userSolutionRepository.saveAll(Flux.just(userScoreDocument1, userScoreDocument2, userScoreDocument3)).blockLast();
+
+        // Para testear ZMQ
+        scoreZmq = 100;
+        errorsZmq = "Error test";
+        userSolutionZmq = UserSolutionDocument.builder()
+                .uuid(UUID.randomUUID())
+                .userId(UUID.randomUUID())
+                .challengeId(UUID.randomUUID())
+                .languageId(UUID.randomUUID())
+                .solutionDocument(solutionDocumentList)
+                .build();
     }
 
     @DisplayName("Repository not null Test")
@@ -259,5 +308,50 @@ public class UserSolutionRepositoryTest {
                 .thenCancel()
                 .verify();
 
+    }
+
+    @DisplayName("Repository not null Test")
+    @Test
+    void testDB(){
+        assertNotNull(userSolutionRepository);
+    }
+
+    @DisplayName("Count users by id")
+    @Test
+    void CountUserById(){
+        Flux<UserSolutionDocument> userScoreDocTest = userSolutionRepository.findByUserId(userId1);
+
+        StepVerifier.create(userScoreDocTest)
+                .expectNextCount(3)
+                .verifyComplete();
+    }
+
+    @DisplayName("Should persist score and errors in existing document when receiving data from ZMQ")
+    @Test
+    void testSaveScoreAndErrorsInExistingDocument() {
+
+        Mono<UserSolutionDocument> savedInitialDocument = userSolutionRepository.save(userSolutionZmq);
+
+        StepVerifier.create(savedInitialDocument)
+                .expectNextMatches(doc -> doc.getScore() == 0 && (doc.getErrors() == null || doc.getErrors().isEmpty()))
+                .verifyComplete();
+
+        Mono<UserSolutionDocument> existingDocument = userSolutionRepository.findByUuid(userSolutionZmq.getUuid());
+
+        Mono<UserSolutionDocument> updatedDocument = existingDocument
+                .map(doc -> {
+                    doc.setScore(scoreZmq);
+                    doc.setErrors(errorsZmq);
+                    return doc;
+                })
+                .flatMap(userSolutionRepository::save);
+
+        StepVerifier.create(updatedDocument)
+                .expectNextMatches(doc -> doc.getScore() == scoreZmq && doc.getErrors().equals(errorsZmq))
+                .verifyComplete();
+
+        StepVerifier.create(userSolutionRepository.findByUuid(userSolutionZmq.getUuid()))
+                .expectNextMatches(doc -> doc.getScore() == scoreZmq && doc.getErrors().equals(errorsZmq))
+                .verifyComplete();
     }
 }
